Extract app mock calls into facade.
